Refactored the generic method to accept multiple path parameters
